clewer kill

+++ Makefile
@@ -5,7 +5,7 @@ install:
 	sudo mkdir -p /home/pi/.config/autostart/
 	sudo cp commando.desktop /home/pi/.config/autostart/commando.desktop
 	sudo cp configs/conf.yaml /etc/commando.yaml
-	pkill commando
+	if pgrep commando; then pkill commando; fi
 	sudo cp bin/commando /usr/bin/commando
 	sudo chmod +x /usr/bin/commando
